my-app/backend/pyproject.toml
[project]
name = "cane-ai-backend"
version = "0.1.0"
description = "Cane AI Incident Response Platform - Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "anthropic>=0.39.0",
    "openai>=1.50.0",
    "cryptography>=42.0.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.9",
    "email-validator>=2.0.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
